+++ CHANGELOG.md
@@ -24,6 +24,7 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 ### :bug: Bug Fixes
 
 * fix(sdk-metrics): fix memory leak in `TemporalMetricProcessor` where unreported accumulations grew unboundedly for slow or absent collectors
+* fix(sdk-trace-web): propagate `optimised` flag in `getElementXPath` recursion [#6335](https://github.com/open-telemetry/opentelemetry-js/pull/6335) @akkupratap323
 
 ### :books: Documentation